#40 implement send notification in background task service
