+++ Infrastructure/BackgroundJobs/Services/BackgroundTaskService.cs
@@ -1,4 +1,5 @@
 ﻿using Application.DTOs.Identity;
+using Application.DTOs.Notification;
 using Application.Interfaces.Services.Hangfire;
 using Infrastructure.BackgroundJobs.Jobs;
 
@@ -14,6 +15,12 @@ namespace Infrastructure.BackgroundJobs.Services
         public Task SendEmailAsync(EmailMessage message)
         {
             _backgroundJobService.Enqueue<EmailJob>(x => x.SendJob(message));
+            return Task.CompletedTask;
+        }
+
+        public Task SendNotificationAsync(NotificationMessage message, string correlationId)
+        {
+            _backgroundJobService.Enqueue<NotificationJob>(x => x.ExecuteAsync(message, correlationId));
             return Task.CompletedTask;
         }
     }
